Finish CRUD for BIll Order

+++ src/main/java/com/example/restaurant_app_spring/service/impl/MenuItemServiceImpl.java
@@ -6,19 +6,21 @@ import com.example.restaurant_app_spring.dto.request.pagination.PaginationReques
 import com.example.restaurant_app_spring.dto.request.sort.SortRequest;
 import com.example.restaurant_app_spring.dto.response.menuitem.MenuItemResponse;
 import com.example.restaurant_app_spring.entity.MenuItem;
-import com.example.restaurant_app_spring.mapper.BillDetailMapper;
-import com.example.restaurant_app_spring.mapper.BillOrderMapper;
+import com.example.restaurant_app_spring.entity.billdetail.BillDetail;
+import com.example.restaurant_app_spring.exception.InvalidDataException;
 import com.example.restaurant_app_spring.mapper.MenuItemMapper;
+import com.example.restaurant_app_spring.repository.BillDetailRepository;
 import com.example.restaurant_app_spring.repository.MenuItemRepository;
 import com.example.restaurant_app_spring.service.MenuItemService;
+import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
+import org.hibernate.Session;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.ExampleMatcher;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import java.util.List;
-import java.util.regex.Matcher;
 
 @Service
 @RequiredArgsConstructor
@@ -26,6 +28,8 @@ public class MenuItemServiceImpl implements MenuItemService {
 
     private final MenuItemRepository menuItemRepository;
     private final MenuItemMapper menuItemMapper;
+    private final BillDetailRepository billDetailRepository;
+
 
     public MenuItemResponse create(MenuItemRequest menuItemRequest) {
         MenuItem menuItem = this.menuItemMapper.convertToEntityFromRequest(menuItemRequest);
@@ -71,6 +75,10 @@ public class MenuItemServiceImpl implements MenuItemService {
     @Override
     public void delete(Integer idToDelete) {
         MenuItem menuItemToDelete = this.menuItemRepository.findById(idToDelete).orElseThrow();
+        List<BillDetail> billDetails = this.billDetailRepository.findBillDetailByMenuItemId(idToDelete);
+        if(billDetails.size() > 0) {
+            throw new InvalidDataException();
+        }
         this.menuItemRepository.delete(menuItemToDelete);
     }
 }
